requirements.txt
asn1crypto==1.5.1
boto3==1.26.146
botocore==1.29.146
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
dash==2.10.2
dash-ace==0.2.1
dash-bootstrap-components==1.4.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
filelock==3.12.0
Flask==2.2.5
halo==0.0.31
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
log-symbols==0.0.14
MarkupSafe==2.1.3
oscrypto==1.3.0
packaging==23.1
plotly==5.14.1
psycopg2==2.9.6
pycparser==2.21
pycryptodomex==3.18.0
PyJWT==2.7.0
pyOpenSSL==23.2.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
s3transfer==0.6.1
six==1.16.0
snowflake-connector-python==3.0.4
sortedcontainers==2.4.0
spinners==0.0.24
tenacity==8.2.2
termcolor==2.3.0
typing_extensions==4.6.3
urllib3==1.26.16
Werkzeug==2.2.3
zipp==3.15.0
